BasicMainCharacterBehavrior should be good now. #149
Former-commit-id: 0f2e3e55b5b5c0fadd682498572454d26d8409f5

+++ src/main/java/ooga/model/entities/maincharacter/MainCharacterBehavior.java
@@ -8,6 +8,6 @@ public interface MainCharacterBehavior extends ImmutableMainCharacterBehavior{
   void checkLivesAndUpdateMainCharacterState();
   void setMainCharacterState(MainCharacterState mainCharacterState);
   void updateScore(double addToScore);
-  int getScore();
+  double getScore();
 
 }
